requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7